sort disk by partition number
I noticed a huge variation on fileio seq write testcase with a same
kernel. Well, the kernel changes indeed due to rebuild, and I thought
that may be the reason; it turned out not. And the real root cause is
a bit interesting and tricky: we used different test partition.

At first time, we used /dev/sda2, and we got a throughput about 100MB/s,
which is the max throughput we can got on that hard disk. We then
reformatted the disk, which results to /dev/sda10(the last parition) as
a test partition. The result suddenly changed to 60MB/s after that,
which sounds like a huge regression, but it's not, actually.

Here sort the disk by partition number so that the faster partition will
be picked up for testing.

v2: fix a the sort pointted out by Wanlong

Signed-off-by: Yuanhan Liu <yuanhan.liu@linux.intel.com>
Signed-off-by: Fengguang Wu <fengguang.wu@intel.com>