fsmark: change the f postfix to fpd
"16d-256f" is kind of clear, but it might confuse people for "1f" only,
which means 1 files_per_diretory actually. This is used for creating as
much directories as files to make the dir and file ratio to 1:1.

Chaning the f postfix to fpd to make it more clear in case like that.

Signed-off-by: Yuanhan Liu <yuanhan.liu@linux.intel.com>
Signed-off-by: Fengguang Wu <fengguang.wu@intel.com>