fsmark: introduce generic fsmark jobs
No small file writes this time(we have fsmark-stree-journal-xx.yaml).
It's just a simple write testcase with write unit set to 4M to test
different device(hdd, ssd, raid, brd), different sync method and
with different threads.

It's a bit like dd-write, and the good thing about fsmark is that it
has a result, so that it makes it clear whether there is regression or not.

Signed-off-by: Yuanhan Liu <yuanhan.liu@linux.intel.com>
Signed-off-by: Fengguang Wu <fengguang.wu@intel.com>