power index: watch turbostat.GHz changes
turbostat.GHz impacts all power/performance numbers. If it changed,
we'd like to show it in compare -ai outputs. If it is bisect-able,
it's worthwhile to bisect down the commit. If it's bisect-able but
bisect failed, it may be due to environment changes -- like temperature,
then we need to fix the temperature to make workloads run more stable.

Signed-off-by: Fengguang Wu <fengguang.wu@intel.com>