Removing rsync of .deb files
My understanding is that "lkp-tests/sbin/pack -d" is to be run by
"lkp-tests/bin/setup_local" only, and within a test box.

So it doesn't make sense for "lkp-tests/sbin/pack -d" to rsync .deb files
to inn server because test boxes may not be able to communicate with inn.

Change-Id: I00b62dbf7efe976f688a733d461a3c87b1639239
Signed-off-by: Helia Correia <helia.correia@intel.com>
Signed-off-by: Fengguang Wu <fengguang.wu@intel.com>