stats/unixbench: catch incomplete result
It's an abnormal condition if no "score". There should be a loud
warning -- not only to prevent silent errors, but also to prevent
silent data corruptions.

If we need to show a warning here anyway, I prefer the original ruby
exception -- the code would be more simple and the error format is
uniform. However the downside is, if stats/unixbench exit abnormally,
the result will be dropped altogether. We may lose the opportunity to
catch some functional failure of unixbench.

So add a warning message and a bisectable stats item.

CC: Yao Weiqi <weiqi.yao@intel.com>
Signed-off-by: Fengguang Wu <fengguang.wu@intel.com>