Revert "fix: require lib/stats.rb in lib/matrix.rb"
This reverts commit e7fd187c17ea7caf0b17a16e1a50671a705bdb80.

1) we should avoid using "require_relative", because "require" can
avoid duplicate including libraries automatically, but will include
twice if the library path is not the same. Says that the relative
path will not be recognized as the same path, then it will be included
twice. To avoid duplicating including libraries, we should always use
"require" with the absolute path.

2) "lib/stats.rb" is not the only required library for "lib/matrix.rb",
in that case, might as well include nothing. And as I can find, the
"matrix.rb" always use with "stats.rb" together, no problem until now.

Signed-off-by: Wanlong Gao <gaowanlong@cn.fujitsu.com>
Signed-off-by: Fengguang Wu <fengguang.wu@intel.com>