job: introduce self-contained job params
It's a bit awkward that we have to modify a params/xxx.yaml every
time we want to add a new combination, especially when there are
millions of combination, say disk, and we just need add another
of them.

It becomes more awkward if we just want to test a new combination
*temporarily*.

Here I propose a solution called self-contained job params to resolve
this issue. As the name already explains, we put the param definition
into the job file, which is idea for temporarily combination.

Example:
    $ cat job.yaml
    testcase: fsmark

    ....

    disk: 12BRD_30G

    fsmark:
      filesize: 4M
      ....

    params:
      12BRD_30G:
        nr_brd: 12
        ramdisk_size: 30G

    --------
    # without this patch
    $ job-env job.yaml | grep disk
    disk: 12BRD_30G

    # with this patch
    $ job-env job.yaml | grep disk
    disk: 12BRD_30G
    disk.nr_brd: 12
    disk.ramdisk_size: 30G

Signed-off-by: Yuanhan Liu <yuanhan.liu@linux.intel.com>
Signed-off-by: Fengguang Wu <fengguang.wu@intel.com>