job-init: global and per job output
Once we support running multiple jobs in without reboot, the output need
to be sperated.

Here I introduced a global output, those copies will be redicted to serial
and console device.

And before start running a job, we make those copies to $TMP, which will
be copied into corresponding RESULT_ROOT when job finished. And hence,
it's a per-job output.

Signed-off-by: Yuanhan Liu <yuanhan.liu@linux.intel.com>
Signed-off-by: Fengguang Wu <fengguang.wu@intel.com>