run-lkp: turn it to a standalone program
So that, hopefully, we could be able to run a (new) job just with:

	run-lkp /path/to/job.yaml

Clearly, it's not good for debug, as we can not run a job twice with
same RESULT_ROOT. jobfile_append_var would be another issue as the job
file at server side is moved.

To make it be friendly for debugging, there are several ways, such as
creating a temp RESULT_ROOT and ignore jobfile_append_var requests.

Practically, we will never run a same job twice. So, no issue so far.
For friendly debug purpose, we could discuss and implement it later(if
it's worthwhile).

Signed-off-by: Yuanhan Liu <yuanhan.liu@linux.intel.com>
Signed-off-by: Fengguang Wu <fengguang.wu@intel.com>