Make locale setting works on non-Debian
It appears that C.UTF-8 is only available on Debian.  Use en_US.UTF-8
for Fedora.

This is to explicitly avoid use of C or POSIX locale.  Because wget will
convert the C or POSIX string to UTF-8 string, during the conversion,
%2B will be convert to +.  We need a UTF-8 locale to prevent conversion.
And UTF-8 is our default encoding use during development, so it is
better to use UTF-8 in our system as much as possible to avoid potential
issues in the future.

Signed-off-by: Huang, Ying <ying.huang@intel.com>
Signed-off-by: Fengguang Wu <fengguang.wu@intel.com>