wakealarm: write relative value to wakealarm
date on my tbox returns wrong time: 08:59 while the correct time is 16:59.
then the wake time calculated by the script is not in the future as the
kernel think so the echo to wakealarm failed, which cause resume from S3
doesn't happen.

Since the wakealarm also accepts relative time, we can use that instead.
It's simpler and clearer. This feature is available since 2.6.26 by
commit c116bc2ae516(rtc: add the support for alarm time relative to
current time in sysfs).

Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Signed-off-by: Fengguang Wu <fengguang.wu@intel.com>