job.sh: fix watchdog wait timeout caused error
run_test() may block to wait for other nodes until reach the timeout
of watchdog, then it should be able to killed by the watchdog.

It can also fix the error:
cat: /tmp/lkp/pid-tests: No such file or directory

This error shows because the real test with wrapper still has no
chance to run, the pid-tests file is not yet created.

Refactor the kill_tests to support the cluster mode. To kill tests
or daemon from child to parent by sequence.

Signed-off-by: Wanlong Gao <gaowanlong@cn.fujitsu.com>
Signed-off-by: Huang, Ying <ying.huang@intel.com>