monitors/latency_stat: do not append trailing new line
We set the buffer mode to "unbuffered" for the output stream in
monitors/wrapper and that would cause the output to go to the terminal
as soon as they appear, which caused trouble for proc files under
/proc/sys where it will return an -EINVAL value for the trailing new
line, as shown here(notice the echo is run twice here):
[root@aaronlu linux]# stdbuf -o0 echo 0 > /proc/sys/kernel/latencytop
[ 1742.129833] [8785->echo:8573->bash] latencytop, count=1, buf: 30, 0
[ 1742.129979] latencytop, count=1, error=1
[ 1742.130048] [8785->echo:8573->bash] latencytop, count=1, buf: a,

[ 1742.130211] latencytop, count=1, error=-22
echo: write error

Such error would occur on any files under /proc/sys, I'll not show more
examples here.

Notice that the first write is successful, so there is no functionality
problem here, just a mis-leading error message.

Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>