git/lib.rb: only set ENV_VARIABLE_NAMES if it is not set yet
I'm using ruby-git 1.2.9.1 and I got a lot of warnings like this:
/home/aaron/src/lkp/lkp-tests/lib/git/lib.rb:35: warning: already initialized constant Git::Lib::ENV_VARIABLE_NAMES
/usr/local/share/gems/gems/git-1.2.9.1/lib/git/lib.rb:843: warning: previous definition of ENV_VARIABLE_NAMES was here

ruby-git 1.2.8, which ships in debian unstable doesn't have this
problem currently but will when the update occurs, so set
ENV_VARIABLE_NAMES only when it's not defined in ruby-git's library
code.

BTW, Fedora ships with a year old 1.2.5 that doesn't work well with
git2(lots of warnings), so use "gem install git" is recommended on Fedora.

Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>