setup/disk: add pmem support
/dev/pmemX is not a partition of the nonexistant disk /dev/pmem, but we
still set $partition to all the available /dev/pmemX so that we can use
it in other scripts such as pmem_setup where we setup swap devices from
the partitions value.
The disk is not useful for pmem, so no need to set it to /dev/pmem,
which doesn't exist anyway.

Signed-off-by: Yanbing Jiang <yanbing.jiang@intel.com>
Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>