ndctl: add missing dependency package libtool
library used but 'RANLIB' is undefined
The usual way to define 'RANLIB' is to add 'AC_PROG_RANLIB'
to 'configure.ac' and run 'autoconf' again.
Libtool library used but 'LIBTOOL' is undefined
The usual way to define 'LIBTOOL' is to add 'LT_INIT'
to 'configure.ac' and run 'aclocal' and 'autoconf' again.
If 'LT_INIT' is in 'configure.ac', make sure
its definition is in aclocal's search path.

Signed-off-by: Xiaolong Ye <xiaolong.ye@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>