lkp-exec/rm-path: add a sanity check
The full logic processing like below:
- revert_unite_rt
- remove_from_mrtts
- remove_index
- remove_paths

due to revert_unite_rt() will call "lkp reprocess", it will lead to sbin/process-result-queue to process the result_root,
and sbin/process-result-queue maybe move the result_root when result_root is "bad".

So need to add a sanity to avoid the latter code to continue to process.

Signed-off-by: Yilong Ren <yilongx.ren@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>