perf-node: do not output metric if it doesn't exist
On Mon, Mar 14, 2016 at 07:35:21PM +0800, Fengguang Wu wrote:
> On Mon, Mar 14, 2016 at 04:34:36PM +0800, Aaron Lu wrote:
> > +if ((sum["node-loads"] + sum["node-load-misses"]) != 0)
> > +	node_local_load_ratio = (100 * sum["node-loads"]) / (sum["node-loads"] + sum["node-load-misses"])
> > +else
> > +	node_local_load_ratio = "N/A"
>
> "N/A" is invalid value for stats. It must be a number.
>
> You may (1) show it as 0, or (2) do not show the line at all.
>

It seems that if I output an invalid value for the key("N/A" instead of a
number), it is also dropped. Maybe this will make an error message
appear in the log? Anyway, I agree that do not output is better, so here
it goes:

From: Aaron Lu <aaron.lu@intel.com>
Subject: [PATCH lkp-tests] perf-node: do not output metric if it doesn't exist

As suggested by Fengguang, if the metric doesn't have a value, do not
show it.

Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>