perf-node: send output to the opened fifo
For monitors, we open a fifo in wrapper for it to dump its data there
and then read the fifo to output the data to $TMP_RESULT_ROOT/$monitor.
The perf-node/perf-stat don't follow this in that it will simply output
the perf's output to a file, which is the same file as where the fifo
goes: $TMP_RESULT_ROOT/$monitor.

This doesn't look quite right and it made any other output in the above
monitors not appear in the final log file. Fix this by sending the
perf's output to the fifo by using --log-fd 1 since wrapper has
redirected the stdout to the fifo.

Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>