perf-profile: update perf used in post-run
The perf-profile generated a perf.data that is processed by perf in
post-run to generate the perf-profile.gz. After we updated the command
used in the monitor, we also have to update the perf used in post-run or
the older version of perf can not parse the newer version saved data.

Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>