jobs: auto enable perf-profile monitor for benchmarks
perf-profile is a heavy weight monitor and we run it for limited
duration to avoid impacting the test workload.

For the jobs with defined runtime, auto run it in the middle of test.
For others let perf-profile start at the default 100s delay which
should be early enough to avoid overrunning the test workload.

The default perf-profile duration is 10 seconds.

CC: Huang Ying <ying.huang@intel.com>
CC: Andi Kleen <ak@linux.intel.com>
Signed-off-by: Fengguang Wu <fengguang.wu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>