filter-job: add support for scheduler filtering a job
It is not correct to filter a job based on the default kconfig files from
either /kbuild-tests/kconfig or /c/kernel-tests/kconfig directories, because
the final kconfig file for building the kernel is not exactly same as the
original one due to the fixup_config process before compiling the kernel.

Thus, we have to rely on the final kconfig file to do the filtering for tests
have dependent on kconfigs. This can NOT be achieved when queuing a job,
because the final kconfig file is not available at that time.

Signed-off-by: Wenzhong Sun <wenzhong.sun@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>