jobs/vm-scalability-swap: add a disk label
The swap setup has to be run after the disk setup since the swap setup
needs the partitions environment variable provided by disk. Fix this by
adding an empty disk label in the common job file as a place holder and
the value will be overwritten by the actual job file.

Suggested-by: Fengguang Wu <fengguang.wu@intel.com>
Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>