post-run: throw out stderr when some monitors are blocked by activate-monitor
To distinguish the cause of empty monitors' log files between monitor program's
issue and sync_cluster_state issue.

sync_cluster_state can abort the test process directly if there is a failure
happened from any node in the cluster. As a result, the test process is not
able to wakeup 'activate-monitor', which causes empty log files for monitors.

This patch is not able to fix the following issue, but reminds the different
cause of empty monitor's log file, so that we can filter out a lot of following
error log when checking lkp logs.
"gzip: /result/.../${monitor_name}.gz: unexpected end of file"

https://jira01.devtools.intel.com/browse/OLT-1137

Acked-by: Fengguang Wu <fengguang.wu@intel.com>
Signed-off-by: Wenzhong Sun <wenzhong.sun@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>