tests/vm-scalability: make it possible to specify unit_size
Most of the vm-scalability test cases use the default unit_case but it
could be useful to let user decide the unit_size, e.g. I'm evaluating
the latencies of swap in path and would like to do a single run of
case-anon-w-rand with size of 124G like this:
usemem -n X --random 124G/X

The following patch will use unit_size in case-anon-w-rand.

Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>