stats/piglit: show pass/fail/block/total_test stats
Now this tool supports a more complete method to show
pass/fail/block/total_test in result files such as stats.json.
Original handling to detailed fail-types are reserved as it was.
Now all test names contain slash '/' are forced to be translated into
dots '.' in output as formal.
It's going to take:
1. 'xxx: pass' as pass
2. 'xxx: fail' as fail, same to other detailed fail-types.
3. 'xxx: skip' as block instead of skip.
4. sum of all above, including detailed fail-types, as total_test

Reviewed-by: Wenzhong Sun <wenzhong.sun@intel.com>
Signed-off-by: Huang, Xunsong <xunsong.huang@intel.com>
Signed-off-by: Fengguang Wu <fengguang.wu@intel.com>