include/perf-profile: use suggest_kconfig option instead of need_kconfig
From Fengguang's suggestion:

For more monitors that depend on a kconfig, we may do a suggest_kconfig:
for them which will have the effect of enabling kconfig
in RHEL kernels, while not causing job file to be deleted.

Monitors are by definition good to have, but not required
for running the test.

Suggested-by: Fengguang Wu <fengguang.wu@intel.com>
Reviewed-by: Fengguang Wu <fengguang.wu@intel.com>
Signed-off-by: Yilong Ren <yilongx.ren@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>