setup: eatmem
Purpose here is to consume $size bytes of memory before the test starts
and till the test is done.

Since the memory consumption has to be done before the test starts,
setup script is used here.

This is achieved by using usemem with detach and sleep(suggested by Ying):
1 the setup script forks usemem and waits for it;
2 usemem fork a process, let's say A, to consume the memory; usemem then
  waits for its child process A to exit;
3 process A forks process B to consume the memory and sleeps on a
  semaphore it creates - this is called detach;
4 process B consumed the memory in the form of anonymous pages. It then
  wakes up process A using the semaphore. Process A will exit after being
  waken up.
5 process B will sleep for X seconds after memory consumption;
6 usemem will exit since its child process A has exited;
7 the setup script is done since usemem is done;
8 the job continues to the next setup script.

Till now, the purpose has achieved: test isn't started and memory
consumption is done by process B. To make sure the memory is consumed
while the test is run, we have to let process B sleep instead of exit.
So the X seconds has to be longer than the test's runtime.

Note that:
1 Memory size to consume must be specified in the job file.

2 X seconds will be 24 hours if not speficied. The pid of the process B
  will be appended to $TMP/pid-monitors so it will be killed upon test
  done(kind of ugly but I do not find another way to do this).

3 One consumer process B is started for each node. Size is distributed
  evenly on all nodes.

Suggested-by: Ying Huang <ying.huang@intel.com>
Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>