jobs/vm-scalability-swap.yaml: set runtime to 300 seconds
All vm-scalability jobs started in LKP are meant to run for 300 seconds.
We do not need set it previouslly because vm-scalability will set a
default runtime but since that has changed, we have to set the runtime
here or the tbox will run for a *very* long time.

$ date
Fri Aug 26 16:38:17 CST 2016

$ lkp wtmp lkp-hsw-ep2
... ...

time: 2016-08-26 15:17:19 +0800
state: running

Reviewed-by: Fengguang Wu <fengguang.wu@intel.com>
Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>