tpcc: add tpcc test case
This test can only run on lkp-bdw-ep2(for now) as it requires some data
which is only available on a lkp-bdw-ep2 attached disk.

The instance_number: 5 is just an example, feel free to use a different
number. The do_fresh means to re-generate the database files and I think
we should always re-generate the database files before the run.

Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>