suspend: save all the numbers for a whole suspend test
In suspend test, a hash table is used to save the time consumption
of each driver/phase.
Previously, the time consumptions are printed out, and the hash table
is deleted for every SINGLE suspend/resume cycle.
But this is not sufficient if we want to get the maximum/minimum
time consumption for a driver/phase during multiple suspend/resume
cycles.

Change the code to save all the time consumption calculated for a
whole suspend test and print them out altogether in the end.

Signed-off-by: Zhang Rui <rui.zhang@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>