stats/wrapper: skip processing the stats for program not in the program_list
This is to solve the problem for cluster test jobs running on the server node
where no program's log file is generated for running server daemons but the
stats processing for the program running on client node will also be handled
on server's result_root due to the current Job2sh algorithm.

In the following example of a netperf test job, the server node will actually
run *netserver* daemon program, but it will still process the *netperf* program's
stats.

if role server
then
	start_daemon $LKP_SRC/daemon/netserver
fi

if role client
then
	run_test test='TCP_SENDFILE' $LKP_SRC/tests/wrapper netperf
fi

extract_stats()
{
	$LKP_SRC/stats/wrapper netperf
	...
}

Signed-off-by: Wenzhong Sun <wenzhong.sun@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>