lib/install.sh: fix shell adapt_packages()
There are only a few packages with architecture specific
annotations right now, they all use a :i386 postfix. The
current ruby adapt_packages() works well, however on shell,
the distro_pkg=${mapping#$pkg:} does not work as well as
Ruby's array use and find. Furthermore the bash implementation
allows a final grep to end up giving you more than one package,
so if you have libc-dev but a distro mapping file has a mapping
for both libc-dev and libc-dev:i386 we'd end up with both
mappings. Lastly the mapping for architectures simply does
not work...

Fix this for now by modifying the shell implementation of
adapt_packages() to check if a package we are going to
query is architecture specific or not, and as a general
rule for now just require non-arch packages to go first,
while arch packages go last. The awk query is much better
and will do what we expected.

Long term we should just modify the distro/adaptation/*
files to use a delimiter which can never be part of a package
name, for instance using the pipe "|". That would allow us to
get rid of this small cruft on the bash implementation and
enable both bash and ruby implementations to match much closely.

Since we are sticking to a rule for now of having the arch
packages go last modify the few archs which do have these
mappings to match this rule.

Signed-off-by: Luis R. Rodriguez <mcgrof@kernel.org>
Reviewed-by: Ye Xiaolong <xiaolong.ye@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>