result: fix tbox containing number is converted to wrong tbox_group
e.g. daixiang-HP-ProBook-450-G3 results in daixiang-HP-ProBook-G3,
which is not expected

This is also related to patch "enable tbox group for vm that has same type and runs on same host tbox group",
that we try to deduce tbox group of vm as "vm-$host_tbox_group-$vm_type".

And if that feature is no longer used, we can revert the patch. But
the fix introduced here is still useful, and align with set_tbox_group
in result.sh.

Reported-by: Xiang Dai <xiangx.dai@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>