lkp-exec/install: fix tbox group config file missing at local
If user has a host named as my-machine-1 which ends with "-"pure numbers,
the host config file generated will be named as my-macheine-1,
but the generated job file will load the host config file named as my-machine.

For instance, this can further result in the issue, that the generated
file can miss the memory setting when running case like linpack.yaml.

Signed-off-by: Xiang Dai <xiangx.dai@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>