bin/setup-local: fix tbox group config file missing at local
If user has a host named as my-machine-1 which ends with "-"pure numbers,
the host config file generated will be named as my-macheine-1,
but the generated job file will load the host config file named as my-machine.
So that the generated file miss the memory setting.
To make sure the memory setting do not miss by generating tbox group config file.

Signed-off-by: Xiang Dai <xiangx.dai@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>