chromeswap: only show throughput in its stat
Since all values in stat file may trigger auto-bisect and eventually
sent out an email to developers, only metric that is proper for auto
bisect should be put into the stat file.

For chromeswap, the only value we should care is the throughput. All
other values are used to aid the analysis of what happened. So filter
them all except the throughput line.

Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>