allot: allocate autotest on physical machine
the tests running on vm are all incomplete run b/c no valid status
file under results dir. This may indciate that vm is not correctly
enabled for autotest or it may mean autotest is not suitable to run
on vm, which need further investigate. To allow monitoring autotest
execution status, use xps as host to run.

Signed-off-by: Philip Li <philip.li@intel.com>