kbuild: use allyesconfig to increase parallellism
'make mrproper', 'make defconfig' and the start & end of the build are
mostly single threaded, to increase overall build parallism, we can
enlarge the 'middle' part of the build. One way to achieve this is to
use the 'allyesconfig' since it enables much more code to be built than
the 'defconfig'.

With this config, on lkp-hsw-ep5, the parallism is now:
	Percent of CPU this job got: 4976%
compared to before:
	Percent of CPU this job got: 3547%

Also change the kbuild stat to produce buildtime_per_iteration:
buildtime_per_iteration: 303.0

Note that lkp-avoton2 is too weak to do a 'allyesconfig' in a reasonable
time, it is terminated after 25 minutes run when it is still busy
building for the 1st iteration. So 'allyesconfig' is not suitable for
machines like lkp-avoton2. Due to this, the new kbuild-fullload.yaml
file is newly created and is assigned to lkp-hsw-ep5 and ivb43 as their
cyclic job while the old kbuild.yaml is still used for lkp-avoton2.

Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>