distro/depends/ndctl: fix unmet dependencies
When install libudev1, the error occurs like below:
Reading state information...
Some packages could not be installed. This may mean that you have
requested an impossible situation or if you are using the unstable
distribution that some required packages have not yet been created
or been moved out of Incoming.
The following information may help to resolve the situation:

The following packages have unmet dependencies:
 udev : Breaks: systemd (< 232-20)
        Breaks: systemd:i386 (< 232-20)

The udev has existed but the version is low, so install it to update.

Signed-off-by: Xiang Dai <xiangx.dai@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>