vm-scalability: remove remap from regular run
The usemem_remap test case uses the remap_file_pages(2) API to reverse
a mapping of a file and then comes out with a throughput.

The problem is, that API is deprecated and set to be removed sometime in
a future kernel release. Quote Documentation/vm/remap_file_pages.txt:
"
The syscall is deprecated and replaced it with an emulation now. The
emulation creates new VMAs instead of nonlinear mappings. It's going to
work slower for rare users of remap_file_pages() but ABI is preserved.
"

It doesn't make much sense to do performance test for it so remove it
from regular runs(user can still manually queue job for it of course).

Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Acked-by: Fengguang Wu <fengguang.wu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>