stats/mpstat: fix time format of mpstat
Use LC_ALL=C to use english to show time in mpstat.
Fix the problem in stats of different time formats.
The reported error message when running jobs with mpstat is

2017-04-18 08:00:50 -0500 WARN – invalid stats value: all

This is caused by the different time format of 12/24 hour system.
Detect time format of 12/24 hours. If the second row is AM or PM but
not 'all', it's 12 hour clock, otherwise it is 24.

Signed-off-by: Wu Chenggang <chenggang.wu@intel.com>
Reviewed-by: "Huang, Ying" <ying.huang@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>