kthread_noise: calculate kthread's runtime to decide if there is noise
kthread is regarded as a noise factor, this monitor calculate how much
time all kthreads have taken during the workload execution time and
produce a total time for comparison between kernels.

This is achieved by getting a list of pids that are kthreads by checking
if /proc/$pid/exe is a valid file: if not, then this is regarded as a
kthread. Then when workload finished, get their execution time from
/proc/$pid/stat field 14 for user time and field 15 for system time. All
kthreads' time are added to produce a final total_time for comparison.

Cc: Ying Huang <ying.huang@intel.com>
Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Acked-by: "Huang, Ying" <ying.huang@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>