pack/piglit: fix an error in function pack
The stdout of `echo /usr/lib/x86_64-linux-gnu/libwaffle-1.so.0.*`
will be on the same line. So it will be considered as one file by
cpio, no matter how many file could match the name.

The error log of pack shows below.
'''''''''''''''''
cpio: /usr/lib/x86_64-linux-gnu/libwaffle-1.so.0.5.2 /usr/lib/x86
_64-linux-gnu/libwaffle-1.so.0.5.90: Cannot stat: No such file or
directory
'''''''''''''''''
To fix it, we use find instead of echo. I have confirmed that it
could work as we want.

Signed-off-by: Hailin Gu <hailinx.gu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>