ftrace-funcgraph: fix hash array generation
I do not quite understand what the problem is, but it doesn't work.
If h = Hash.new([]), then:
   h['A'] << 1
will give us
h['A'] = [1], h['B'] = [1], h['C'] = [1], etc...

This is not what we want, fix it.
BTW, I also do not understand the = sign before <<, but removing it
doesn't cause any problem, so it's removed.

Cc: Ying Huang <ying.huang@intel.com>
Acked-by: "Huang, Ying" <ying.huang@intel.com>
Signed-of-by: Aaron Lu <aaron.lu@intel.com>

Signed-off-by: Philip Li <philip.li@intel.com>