exit_free: rename time to duration
time is special in lib/stats.rb - it will ignore metric with this name.
Rename time to duration to work around it.

BTW, this metric seems redundant, since all the test did is to wait and
we can have time information by adding "category: benchmark" to its job
file. Anyway, better to output something than nothing so keep it.

Cc: Ying Huang <ying.huang@intel.com>
Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Acked-by: "Huang, Ying" <ying.huang@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>