distro/depends/xfstests: remove xfsprogs
Now we install xfsprogs from repo(the version is v4.12) so do not need install it from deb source.
Also, since xfsdump depends on xfsprogs, we remove it in case install it would
override what we install from repo.
Only one testcase need xfsdump is xfs/022, which has ignored. If need enable it in
future, please also install it from repo or ensure the version of xfsprogs is over v4.10.

Signed-off-by: Xiang Dai <xiangx.dai@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>