backup kexec-free solution for testboxes without PXE ROM
the iPXE service helps us deploy a testbox without using kexec.
But iPXE service requires native PXE ROM, or else, even if iPXE kernel is
loaded from local disk, there is still no way for iPXE kernel to download
Job kernel from PXE server, because there is no ethernet support.

Previously, we use a USB key with kexec kernel installed to load and kexec to
the Job kernel, for testboxes without PXE ROM.

Now, an alternative solution is introduced in this patch, with the help of
grub, we can still deploy the testbox without using kexec.

The main idea of this patch is to leverage grub-reboot command, which allows
the system to boot to a specified menuentry after reboot, in one shot mode,
and then switch back to the default menuentry after another reboot.

So for any testbox without PXE ROM, we can install Linux either on local disk,
or on a USB key, and override the /etc/rc.local script with this one, and then
everything should work as expected.

The work flow is like below
1. testbox boots to default kernel either from local disk or from USB key
2. network is setup by the default kernel
3. rc.local download Job kernel image and initrd image from LKP WEB server
4. rc.local run grub-reboot "Job kernel" and update-grub
5. Job kernel is loaded and booted from local disk/USB key
6. After job finished, system reboots to default kernel again, jump to step 1

Signed-off-by: Zhang Rui <rui.zhang@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>