pack/mdadm-selftests: fix test 10ddf-create
10ddf-create try to get the device number, rootfs is not supported,
let it to read the ext3 filesystem created from an external hdd.

Rootfs is a special instance of ramfs (or tmpfs, if that's enabled) as a dynamically
resizable RAM-based filesystem, We should use a persistent filesystem (e.g. ext3 ext4)
instead of a temporary filesystem (e.g. rootfs) to enable the test.

------------
$ sudo mount -l | grep ext4
/dev/mapper/ubuntu--vg-root on / type ext4 (rw,relatime,errors=remount-ro,data=ordered)
$ sudo ./test --tests=10ddf-create
Testing on linux-4.2.0-42-generic kernel
tests/10ddf-create... succeeded
------------

Signed-off-by: Chen Rong <rongx.a.chen@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>