pft: remove pft.patch
The patch essentially removed NUMA functionality,
since we need NUMA functionality, we delete it now.

It's not clear why the patch is added in the first place, a guess is
that initially, pft is not meant to be tested on NUMA machines and
to avoid building with libnuma, the stub is created.

Cc: Ying Huang <ying.huang@intel.com>
Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>