plot: fallback Gnuplot.open() to version v2.4.1 one
The Gnuplot.open() defined in ruby_gnuplot lib v2.6.2 creates circular
dependency and may cause potential dead lock.

The method defined in v2.6.2 is as below:

         def Gnuplot.open( persist=true  )
           cmd = Gnuplot.gnuplot( persist  )
           IO::popen( cmd, "w+" ) { |io|
=>           yield io
             io.close_write
=>           @output = io.read
            }
           return @output
         end

The IO::popen is called with 'w+' which would create two pipes, the work flow is
like:
				io.read
         -----------------------<----------------------------------
         |                                                        |
         |                ______                        ______    |
    ruby process --->---- )_____)--->--- gnuplot --->---)_____) ---
                           pipe1                         pipe2

In this case, ruby process and gnuplot would run in parallel, once gnuplot gets
a chance to run and generate enough output to make the pipe2 full, while ruby
process hasn't finished writing data to gnuplot, gnuplot would be blocked
because ruby process is still writing and won't read pipe2 (to clear its
buffer), then ruby process would be blocked too because gnuplot is blocked and
doesn't get a chance to read pipe1.

IO::popen in v2.4.1 is called with "w", so it will only have 1 pipe which
avoids the circular dependency.

Cc: Fengguang Wu <fengguang.wu@intel.com>
Signed-off-by: Xiaolong Ye <xiaolong.ye@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>