ltp: clear dir_nlink feature for rename11 on ext4
rename11:3 needs the two following conditions to run properly:
1)the file system has a subdirs limit;
2)the subdirs limit can be reached.
Now lkp tests ext4, xfs and btrfs, and only ext4 has a subdirs
limit as 65000. dir_nlink is a feature for ext4 that allows
more than 65000 subdirs per directory and it is default added
when we make file system without specific parameters. In kernels
before v4.13-rc4, even we clear dir_nlink feature in ext4, it
will be automatically added when we reach the suddirs limit
which leads to the failure of getting the max number of subdirs.
So for ltplite test on ext4, we clear dir_nlink to make it work
properly for kernels v4.13-rc4 and later.

fix the following issue:
----------
rename11    0  TINFO  :  Failed reach the subdirs limit
on EXT2/EXT3/EXT4 filesystem
rename11    3  TCONF  :  rename11.c:175: EMLINK test is
not appropriate
----------

Signed-off-by: Fei Jie <feij.fnst@cn.fujitsu.com>
Signed-off-by: Philip Li <philip.li@intel.com>