fs: move unount_fs to post-run.fs
There is a potential bug in setup/fs, refer to the issue of fail to run
postmark benchmark, reported by Philip.

Consider to run postmark.yaml jobfile. In job.sh, there are
      a) run_setup fs='btrfs' $LKP_SRC/setup/fs
      b) run_setup fs='nfsv4' $LKP_SRC/setup/fs1

When runing setup/fs at the first time, the hdd/ssd partition is mounted
via a call to mount_fs(). When running setup/fs1 (fs1 links to fs) at the
second time, the hdd partition is unmounted via a call to unmount_fs()
which is a clear operation at the beginning of running setup/fs script.
Thus, this unmeant operation unmounts the hdd/ssd partition that will be
used to mount nfs later, and therefore leads to failure of mount.nfs.

The proposal is to move this clear operation to post-run.fs, as suggested by
Ying Huang.

Changelog:
    Since V2
    a) bug fix: replace $mt by \$mt since we want to keep the $ symbol
    itself.

    Since V1
    a) move umount_fs from bin/post-run to post-run.fs

Suggested-by: Ying Huang <ying.huang@intel.com>
Signed-off-by: Kemi Wang <kemi.wang@intel.com>
Reviewed-by: "Huang, Ying" <ying.huang@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>