autotest: fix tsc.fail: set the maximum number of CPU that can be used
fix the following issue:
-----------
autotest/tsc.fail:TestFail: Latency CPU 7 - CPU 24 = 763 exceeds threshold 650
-----------
Checktsc is a user space program that checks TSC synchronization between pairs of CPUs on an SMP system.
It will check the tsc between 2 cpus whether <= a given value(here is 650 cycles).
But the result will be unstable, especially under lots number cpus platform since it depends on the runtime environment, scheduler policy etc.
When the number of CPUs is large(eg: 32)，the diference between CPUs is easily larger than the threshold value.
So, we set the maximum number is 16.

Signed-off-by: leishaoting <leist.fnst@cn.fujitsu.com>
Signed-off-by: Philip Li <philip.li@intel.com>