analyze_suspend: resolve OOM issues
We ran into OOM problem when running analyze_suspend on the lkp tboxes.
And the root cause is that
analyze_suspend does the following things
1. enable ftrace
2. capture the dmesg and ftrace logs as result files
3. process the dmesg and ftrace logs to generate html result file

Usually, we only need memory in step 1 to enable ftrace,
and in step 3 to process.
But in LKP, where the ramdisk is used, we need extra memory in step 2
to save the dmesg and ftrace logs, and also extra memory in step 3
to save the html output file.
And these files may be large when ftrace is enabled (hundreds of MB or larger).

All of the above memory usages result in OOM when running analyze_suspend
in LKP. To solve this, analyze_suspend
a) introduces -gzip parameter to gzip the trace and dmesg logs to save space
b) introduces -skiphtml parameter to skip generating html file
c) introduces -bufsize parameter to allow user specifying the buffer size
   took by ftrace, to make sure we can leave enough memory for step 2 and 3.

This patch leverages the above parameters to get rid of OOM in LKP.

Signed-off-by: Zhang Rui <rui.zhang@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>